# Improve function arguments parameter expression detection