Currently we don't detect expressions properly which causes extra environment to be created even though it isn't necessary. Requires more AST analysis.